REF: servicemap e2e
